+++ idf_component_manager/utils.py
@@ -5,31 +5,39 @@ import re
 import click
 from packaging import version
 
-from idf_component_tools.errors import UserHint
+from idf_component_tools.messages import UserHint, UserNotice
 
 CLICK_SUPPORTS_SHOW_DEFAULT = version.parse(click.__version__) >= version.parse('7.1.0')
 
 
-def print_stderr_prefixed(prefix, color, message):  # type: (str, str, Exception | str) -> None
+def print_prefixed(
+    prefix, color, message, stderr=True
+):  # type: (str, str, Exception | str, bool) -> None
     styled_prefix = click.style('{}: '.format(prefix), fg=color)
-    click.echo(styled_prefix + str(message), err=True)
+    click.echo(styled_prefix + str(message), err=stderr)
 
 
 def print_error(message):  # type: (Exception | str) -> None
-    print_stderr_prefixed('ERROR', 'red', message)
+    print_prefixed('ERROR', 'red', message)
 
 
 def print_warn(message):  # type: (Exception | str) -> None
-    print_stderr_prefixed('WARNING', 'yellow', message)
+    print_prefixed('WARNING', 'yellow', message)
 
 
 def print_hint(message):  # type: (Exception | str) -> None
-    print_stderr_prefixed('HINT', 'yellow', message)
+    print_prefixed('HINT', 'yellow', message)
+
+
+def print_notice(message):  # type: (Exception | str) -> None
+    print_prefixed('NOTICE', 'green', message, stderr=False)
 
 
 def showwarning(message, category, filename, lineno, file=None, line=None):
     if category is UserHint:
         print_hint(message)
+    elif category is UserNotice:
+        print_notice(message)
     else:
         print_warn(message)