+++ idf_component_tools/manifest/__init__.py
@@ -15,7 +15,10 @@ from .manifest import (
     ComponentWithVersions,
     HashedComponentVersion,
     Manifest,
+    OptionalDependency,
+    OptionalRequirement,
     ProjectRequirements,
+    filter_optional_dependencies,
 )
 from .schemas import BUILD_METADATA_KEYS, INFO_METADATA_KEYS, JSON_SCHEMA
 from .solved_component import SolvedComponent
@@ -23,22 +26,25 @@ from .solved_manifest import SolvedManifest
 from .validator import ManifestValidator
 
 __all__ = [
+    'BUILD_METADATA_KEYS',
     'ComponentRequirement',
     'ComponentVersion',
     'ComponentWithVersions',
+    'filter_optional_dependencies',
     'FULL_SLUG_REGEX',
     'HashedComponentVersion',
+    'INFO_METADATA_KEYS',
+    'JSON_SCHEMA',
+    'known_targets',
     'MANIFEST_FILENAME',
     'Manifest',
     'ManifestManager',
     'ManifestValidator',
+    'OptionalDependency',
+    'OptionalRequirement',
     'ProjectRequirements',
     'SLUG_REGEX',
     'SolvedComponent',
     'SolvedManifest',
     'WEB_DEPENDENCY_REGEX',
-    'JSON_SCHEMA',
-    'BUILD_METADATA_KEYS',
-    'INFO_METADATA_KEYS',
-    'known_targets',
 ]