+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.4.1] - 2023-10-02
+
+### Fixed
+
+- Fix typo in gitlab CI user agent
+- Fix issue with creating a profile in `idf_component_manager.yml` if it does not exist, when executing `compote registry login` command with `--service-profile` specified
+- Remove warnings of the unknown root keys in the manifest files
+- Fix schema validation of `idf_component_manager.yml`, for an empty profile name value
+- Fix manifest schema validation message for an empty field value
+- Fix processing of `rules` and `matches` for components from the registry
+- Fix processing handling of `require` field for components from the registry
+
 ## [1.4.0] - 2023-09-15
 
 ### Added